+++ testers/instrument_checkout/__init__.py
@@ -1,0 +1,1 @@
+"""Bench instrument checkout: BK 4052 function generator and BK 5492B DMM."""