+++ testers/latching_board_tests/__init__.py
@@ -1,1 +1,0 @@
-"""ePBR27 latching board tester."""